unified-platform/backend/requirements.txt
# Core Django + Wagtail
django>=5.0,<6.0
wagtail>=6.0,<7.0

# Database
psycopg2-binary>=2.9
dj-database-url>=2.1

# REST API
djangorestframework>=3.14
django-cors-headers>=4.3

# Async tasks
celery>=5.3
redis>=5.0

# Server
gunicorn>=21.2
# Utilities
django-environ>=0.11

# PDF Processing
pdfplumber>=0.11.0

# Google AI
google-generativeai>=0.3.2

# US Address Fields
django-localflavor>=4.0
django-phonenumber-field[phonenumbers]>=7.0
